+++ runtime/onert/api/nnfw/src/nnfw_internal.cc
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-#include "nnfw_api_internal.h"
+#include "nnfw_internal.h"
+
+#include "nnfw_session.h"
 
 #include <util/ConfigSource.h>
 
@@ -35,4 +37,22 @@ NNFW_STATUS nnfw_get_config(nnfw_session *session, const char *key, char *value,
 {
   NNFW_RETURN_ERROR_IF_NULL(session);
   return session->get_config(key, value, value_size);
+}
+
+NNFW_STATUS nnfw_load_circle_from_buffer(nnfw_session *session, uint8_t *buffer, size_t size)
+{
+  NNFW_RETURN_ERROR_IF_NULL(session);
+  return session->load_circle_from_buffer(buffer, size);
+}
+
+NNFW_STATUS nnfw_load_model_from_modelfile(nnfw_session *session, const char *file_path)
+{
+  NNFW_RETURN_ERROR_IF_NULL(session);
+  return session->load_model_from_modelfile(file_path);
+}
+
+NNFW_STATUS nnfw_train_export_circleplus(nnfw_session *session, const char *path)
+{
+  NNFW_RETURN_ERROR_IF_NULL(session);
+  return session->train_export_circleplus(path);
 }
