+++ runtime/onert/api/nnfw/src/nnfw_session.h
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-#ifndef __API_NNFW_API_INTERNAL_H__
-#define __API_NNFW_API_INTERNAL_H__
+#ifndef __API_NNFW_SESSION_H__
+#define __API_NNFW_SESSION_H__
 
 #include "nnfw.h"
-#include "nnfw_experimental.h"
 
 #include "CustomKernelRegistry.h"
 #include "compiler/CompilerOptions.h"
@@ -231,4 +230,4 @@ private:
   std::filesystem::path _model_path;
 };
 
-#endif // __API_NNFW_API_INTERNAL_H__
+#endif // __API_NNFW_SESSION_H__
